add friend accept logic

+++ Server/UserEntity.hpp
@@ -4,7 +4,7 @@
 class UserEntity
 {
 private:
-    uint64_t        m_Id;       // 유저 식별키
+    uint32_t        m_Id;       // 유저 식별키
     std::string     m_UserId;   // 유저 아이디
     std::string     m_Password; // 유저 패스워드
     std::string     m_Username; // 유저 이름
@@ -14,7 +14,7 @@ private:
 public:
 
     // Getter methods
-    uint64_t GetId() const { return m_Id; }
+    uint32_t GetId() const { return m_Id; }
     std::string GetUserId() const { return m_UserId; }
     std::string GetPassword() const { return m_Password; }
     std::string GetUsername() const { return m_Username; }
@@ -22,7 +22,7 @@ public:
     char GetIsAlive() const { return m_IsAlive; }
 
     // Setter methods
-    void SetId(uint64_t id) { m_Id = id; }
+    void SetId(uint32_t id) { m_Id = id; }
     void SetUserId(const std::string& userId) { m_UserId = userId; }
     void SetPassword(const std::string& password) { m_Password = password; }
     void SetUsername(const std::string& username) { m_Username = username; }
